Move DiagnosticComparer
