+++ analyzers/tests/SonarAnalyzer.TestFramework.Test/Verification/IssueValidation/DiagnosticsComparerTest.cs
@@ -22,7 +22,7 @@ using SonarAnalyzer.TestFramework.Verification.IssueValidation;
 using CS = SonarAnalyzer.Rules.CSharp;
 using VB = SonarAnalyzer.Rules.VisualBasic;
 
-namespace SonarAnalyzer.Test.Helpers
+namespace SonarAnalyzer.TestFramework.Test.Verification.IssueValidation
 {
     [TestClass]
     public class DiagnosticsComparerTest
